RavenDB-20318 refactoring /task and /tasks
