+++ src/Raven.Server/Web/System/OngoingTasksHandler.cs
@@ -126,7 +126,7 @@ namespace Raven.Server.Web.System
         [RavenAction("/databases/*/task", "GET", AuthorizationStatus.ValidUser, EndpointType.Read)]
         public async Task GetOngoingTaskInfo()
         {
-            using (var processor = new OngoingTasksHandlerProcessorForGetOngoingTaskInfo(this))
+            using (var processor = new OngoingTasksHandlerProcessorForGetOngoingTask(this))
                 await processor.ExecuteAsync();
         }
